+++ packages/web/navigation/SidebarItem.tsx
@@ -10,9 +10,6 @@ import { LinkProps, TextHeadline } from '../typography';
 
 import { useSidebarContext } from './SidebarContext';
 
-/**
- * @deprecated this component will be removed from cds-web Q22023. It has been moved to cds-web-overlays.
- */
 export type SidebarItemProps = {
   /**
    * The Navigation Icon this item represents
@@ -57,9 +54,6 @@ export type SidebarItemProps = {
   Pick<React.AllHTMLAttributes<Element>, 'target'> &
   Pick<LinkProps, 'openInNewWindow'>;
 
-/**
- * @deprecated this component will be removed from cds-web Q22023. It has been moved to cds-web-overlays.
- */
 export const SidebarItem = memo(
   forwardRef(
     (